Extract corridor solver into marine_nav_utilities (#59)
Move the pure corridor solver (CorridorParams, Station, makeLateralOffsets,
solveCorridorOffsets, resampleStations) and its unit tests out of
marine_nav_behavior_tree into marine_nav_utilities/corridor_solver.{h,cpp} so
the controller-layer avoider (PR2) can share it. The BT node re-exports the
symbols via using-declarations — no behavior change. applyAvoidanceSlowdown and
its tests stay with the node (removed in PR2).

PR1 of the #59 controller-decorator rework. Build green;
test_corridor_solver 10/10, test_adjust_path_for_obstacles 2/2.
